+++ demos/svelte/CHANGELOG.md
@@ -1,5 +1,12 @@
 # demos-svelte
 
+## 0.0.1-alpha.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @burnt-labs/abstraxion-js@1.0.0-alpha.2
+
 ## 0.0.1-alpha.0
 
 ### Patch Changes